requirements.txt
python-telegram-bot[job-queue]==20.7
requests==2.31.0
python-dotenv==1.0.0
asyncpg==0.29.0
beautifulsoup4==4.12.2
lxml==4.9.3
apscheduler==3.10.4
psutil==5.9.6
aiohttp==3.9.1
cbrapi==0.1.5